+++ lib/inc/lqr_pid_control.h
@@ -4,23 +4,23 @@
 
 typedef struct pid_parameter
 {
-    float p_gain;
-    float i_gain;
-    float d_gain;
-    float target;
-    float error;
-    float pri_error;
-    float error_limit;
-    float error_sum;
-    float error_sum_limit;
-    float output;
+    double p_gain;
+    double i_gain;
+    double d_gain;
+    double target;
+    double error;
+    double pri_error;
+    double error_limit;
+    double error_sum;
+    double error_sum_limit;
+    double output;
 } pid_parameter_t;
 
 class pid_controller
 {
     public:
         pid_controller();
-        pid_controller(float p_gain, float i_gain, float d_gain)
+        pid_controller(double p_gain, double i_gain, double d_gain)
         {
             this->parameter.p_gain = p_gain;
             this->parameter.i_gain = i_gain;
@@ -30,9 +30,9 @@ class pid_controller
         {
             this->parameter = parameter;
         }
-        void init(float p_gain, float i_gain, float d_gain);
-        void set_target(float target);
-        float calculate(float current_value);
+        void init(double p_gain, double i_gain, double d_gain);
+        void set_target(double target);
+        double calculate(double current_value);
 
     private:
         pid_parameter_t parameter;
