+++ lib/src/lqr_pid_control.cpp
@@ -11,7 +11,7 @@ pid_controller::pid_controller()
     this->parameter.output = 0;
 }
 
-void pid_controller::init(float p_gain, float i_gain, float d_gain)
+void pid_controller::init(double p_gain, double i_gain, double d_gain)
 {
     this->parameter.p_gain = p_gain;
     this->parameter.i_gain = i_gain;
@@ -21,12 +21,12 @@ void pid_controller::init(float p_gain, float i_gain, float d_gain)
     this->parameter.output = 0;
 }
 
-void pid_controller::set_target(float target)
+void pid_controller::set_target(double target)
 {
     this->parameter.target = target;
 }
 
-float pid_controller::calculate(float current_value)
+double pid_controller::calculate(double current_value)
 {
     this->parameter.error = this->parameter.target - current_value;
